requirements.txt
pyserial==3.5
scapy==2.6.0
